+++ R/appMetadata.R
@@ -67,6 +67,14 @@ appMetadata <- function(appDir,
       appDir = appDir,
       appPrimaryDoc = appPrimaryDoc
     )
+    if (appMode == "quarto-shiny") {
+      if (!"knitr" %in% quartoInfo[["engines"]]) {
+        cli::cli_abort(c(
+          "The Quarto document requires a server but does not use an executable engine.",
+          "Consider including some executable code, specifying an engine, or removing the server configuration."
+        ))
+      }
+    }
   } else {
     quartoInfo <- NULL
   }